fix(OMN-E0): resolve validator violations in delegation dispatch port
- Extract _ProtocolDispatchPort to runtime/protocols/protocol_delegation_dispatch_port.py,
  renamed to ProtocolDelegationDispatchPort (PascalCase, no leading underscore)
- Removes mixed models+protocols in service_delegation_dispatch_port.py (architecture gate)
- Eliminates redundant # type: ignore[union-attr] — mypy is clean against typed Protocol
- Add ProtocolDelegationDispatchPort to KNOWN_INFRA_PROTOCOLS allowlist ([RUNTIME] category)
- Run ruff format on handler_wiring.py
- Stamp schema fingerprint (69 migration files)
